+++ CHANGES.md
@@ -28,6 +28,8 @@
 * #103 Allow to define MSP messages with ID > 200 that do not acknowledge
 * #104 Vertical velocity control when in Altitude Hold mode and outside of the deadband 
 * #105 Filter ESKF estimated linear velocities with a LPF
+* #116 Enable immediate execution of actions
+* #116 When executing a flight mission, switching the value of the aux 1 in the remote returns the control of the drone to the TX or APP
 
 ### Changed